Rename snapshot test files to seperate sketch snapshot test files
As a result of the changes made in the index.js file of the mapper
a beforeAll() method was added that initializes the mapper
with the correct file extension.
