+++ lib/__tests__/sketch.test.js
@@ -1,5 +1,5 @@
 const { getColorId, getColorsFromArtboard, getPageArrays } = require('../sketch');
-const dump = require('../../__mocks__/sample_dump.json');
+const dump = require('../../__mocks__/sample_dump_sketch.json');
 
 describe('[sketch.js] test suite', () => {
   describe('-> #getColorId', () => {
