+++ lib/mappers/__tests__/__fixtures__/mapper.sketch.js
@@ -1,4 +1,4 @@
-const sampleOutput = require('../../../../__mocks__/sample_dump.json');
+const sampleOutput = require('../../../../__mocks__/sample_dump_sketch.json');
 const {
   getPageArrays, getGradientsFromArtboard, getColorsFromArtboard, getShadowsFromArtboard,
   getBordersFromArtboard,
