+++ lib/mappers/__tests__/mapper.sketch.test.js
@@ -1,7 +1,12 @@
-const FIXTURES = require('./__fixtures__/mapper');
+const FIXTURES = require('./__fixtures__/mapper.sketch');
 
-const mapper = require('..');
-const formatColor = require('../color');
+const getMapper = require('..');
+const formatColor = require('../sketch/color');
+
+let mapper;
+beforeAll(() => {
+   mapper = getMapper("sketch");
+})
 
 describe('[mapper.js] test suite', () => {
   describe('-> #formatColor:', () => {
